Fix evaluation coefficient data type issues.

+++ research-calculations-java/src/edu/kopp/phd/express/standards/DataFlowValidator.java
@@ -6,7 +6,7 @@ import edu.kopp.phd.express.standards.api.Validator;
 public class DataFlowValidator implements Validator {
 
     @Override
-    public int validate(Model model) {
+    public double validate(Model model) {
         return (countFunctions(model) - countValidDFFunctions(model)) +
                 (countDataStores(model) - countValidDataStores(model)) +
                 (countExternalEntities(model) - countValidExternalEntities(model)) +
